Refactor test_gdq into test_parameter_generation_utils
